requirements.txt
requests>=2.25.1
Pillow>=8.0.0
tkinter  # Python内置，无需安装